slang-bridge/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
schedule==1.2.0

